# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Deel Backend Task API
  version: 1.0.0
  description: Auto-generated OpenAPI documentation for the Deel backend task
paths:
  /admin/best-profession:
    get:
      summary: Get the profession that earned the most money in a time range
      description: >-
        Returns the profession with the highest total earnings between the given
        start and end dates.
      tags:
        - Admin
      parameters:
        - in: query
          name: start
          required: true
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - in: query
          name: end
          required: true
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Best profession found
          content:
            application/json:
              schema:
                type: object
                properties:
                  profession:
                    type: string
                    example: Software Developer
                  earnings:
                    type: number
                    format: float
                    example: 20000
        '400':
          description: Invalid date range or missing parameters
        '500':
          description: Internal server error
  /admin/best-clients:
    get:
      summary: Get the clients that paid the most in a time range
      description: >-
        Returns the top clients ranked by their total payments, limited by the
        optional `limit` query param.
      tags:
        - Admin
      parameters:
        - in: query
          name: start
          required: true
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - in: query
          name: end
          required: true
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - in: query
          name: limit
          required: false
          description: Max number of clients to return (default is 2)
          schema:
            type: integer
            default: 2
      responses:
        '200':
          description: Top clients retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 12
                    fullName:
                      type: string
                      example: John Doe
                    paid:
                      type: number
                      example: 30500
        '400':
          description: Invalid parameters
        '500':
          description: Internal server error
...